fix as per review comments

+++ test-clients/src/main/java/com/hedera/services/bdd/suites/crypto/CryptoCreateForSuiteRunner.java
@@ -50,8 +50,8 @@ public class CryptoCreateForSuiteRunner extends HapiApiSuite {
 	private String nodes;
 	private String defaultNode;
 
-	//Use more initialBalance for this account as it is used as payer for the performance tests
-	private static double initialBalance = 5.0 * (LoadTest.initialBalance.getAsLong());
+	// Use more initialBalance for this account as it is used as payer for the performance tests
+	private static long initialBalance = (long) (5.0 * LoadTest.initialBalance.getAsLong());
 
 	public CryptoCreateForSuiteRunner(String nodes, String defaultNode) {
 		this.nodes = nodes;
@@ -80,7 +80,7 @@ public class CryptoCreateForSuiteRunner extends HapiApiSuite {
 									while (!createdAuditablePayer) {
 										try {
 											var cryptoCreateOp = cryptoCreate("payerAccount")
-													.balance((long) initialBalance)
+													.balance(initialBalance)
 													.withRecharging()
 													.rechargeWindow(3)
 													.key(GENESIS)
